# openapi.yaml
openapi: 3.0.3
info:
  title: HabitOS API
  version: 1.0.0
  description: API for HabitOS - Behavioral Optimization Platform
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - timestamp

    # Auth Schemas
...
    BehaviorCategory:
      type: string
      enum: [health, productivity, learning, social, financial, creativity, mindfulness]
    TimeSlot:
      type: string
      enum: [early_morning, morning, midday, afternoon, evening, night]
...
    ObjectiveImpact:
      type: object
      properties:
        objectiveId:
          type: string
        objectiveName:
          type: string
        impactScore:
          type: number
          description: Score from -1 to 1
      required:
        - objectiveId
        - objectiveName
        - impactScore
    Behavior:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/BehaviorCategory'
        energyCost:
          type: integer
          minimum: 1
          maximum: 10
        durationMin:
          type: integer
        durationMax:
          type: integer
        preferredTimeSlots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
        objectiveImpacts:
          type: array
          items:
            $ref: '#/components/schemas/ObjectiveImpact'
        isActive:
          type: boolean
        frequency:
          type: string
          enum: [daily, weekly, custom]
        frequencyCount:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - description
        - category
        - energyCost
        - durationMin
        - durationMax
        - preferredTimeSlots
        - objectiveImpacts
        - isActive
        - frequency
        - createdAt
        - updatedAt
...
    OptimizationStatus:
      type: string
      enum: [pending, running, completed, failed]
    SolverStatus:
      type: string
      enum: [optimal, feasible, infeasible, unbounded, timeout]
    ScheduledBehavior:
      type: object
      properties:
        id:
          type: string
        behaviorId:
          type: string
        behavior:
          $ref: '#/components/schemas/Behavior'
        scheduledDate:
          type: string
          format: date
        timeSlot:
          $ref: '#/components/schemas/TimeSlot'
        startTime:
          type: string
          example: "08:00"
        endTime:
          type: string
          example: "09:00"
        duration:
          type: integer
        isCompleted:
          type: boolean
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - behaviorId
        - behavior
        - scheduledDate
        - timeSlot
        - startTime
        - endTime
        - duration
        - isCompleted
    ObjectiveContribution:
      type: object
      properties:
        objectiveId:
          type: string
        objectiveName:
          type: string
        contribution:
          type: number
        percentage:
          type: number
      required:
        - objectiveId
        - objectiveName
        - contribution
        - percentage
    OptimizationRun:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        status:
          $ref: '#/components/schemas/OptimizationStatus'
        solverStatus:
          $ref: '#/components/schemas/SolverStatus'
        scheduledBehaviors:
          type: array
          items:
            $ref: '#/components/schemas/ScheduledBehavior'
        objectiveContributions:
          type: array
          items:
            $ref: '#/components/schemas/ObjectiveContribution'
        totalScore:
          type: number
        executionTimeMs:
          type: integer
        constraintsSatisfied:
          type: integer
        constraintsTotal:
          type: integer
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - userId
        - status
        - scheduledBehaviors
        - objectiveContributions
        - totalScore
        - executionTimeMs
        - constraintsSatisfied
        - constraintsTotal
        - createdAt
    OptimizationRequest:
      type: object
      properties:
        targetDate:
          type: string
          format: date
        includeInactiveBehaviors:
          type: boolean
        maxExecutionTimeMs:
          type: integer
    OptimizationResult:
      type: object
      properties:
        run:
          $ref: '#/components/schemas/OptimizationRun'
        schedule:
          $ref: '#/components/schemas/DailySchedule'
      required:
        - run
        - schedule
    DailySchedule:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        date:
          type: string
          format: date
        scheduledBehaviors:
          type: array
          items:
            $ref: '#/components/schemas/ScheduledBehavior'
        totalDuration:
          type: integer
        totalEnergySpent:
          type: integer
        objectiveScores:
          type: array
          items:
            $ref: '#/components/schemas/ObjectiveContribution'
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - date
        - scheduledBehaviors
        - totalDuration
        - totalEnergySpent
        - objectiveScores
        - createdAt

    # Analytics Schemas
...
paths:
...
  /optimization/solve:
    post:
      summary: Run optimization
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizationRequest'
      responses:
        '200':
          description: Optimization result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OptimizationResult'
  /optimization/history:
    get:
      summary: Get optimization history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Optimization history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/OptimizationRun'
...
  /schedule:
    get:
      summary: Get daily schedule
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
          required: false
      responses:
        '200':
          description: Daily schedule
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DailySchedule'
  /schedule/{scheduledBehaviorId}/complete:
    post:
      summary: Mark behavior as complete
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: scheduledBehaviorId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Marked as complete
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
  /schedule/{scheduledBehaviorId}/incomplete:
    post:
      summary: Mark behavior as incomplete
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: scheduledBehaviorId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Marked as incomplete
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string

  # Analytics
...